requirements.txt
neo4j~=5.20
python-dotenv~=1.0
scikit-learn~=1.3
rich~=13.7
streamlit~=1.28.0

